requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
google-generativeai==0.3.2
pandas==2.1.4
openpyxl==3.1.2
Pillow==10.1.0
PyPDF2==3.0.1
Werkzeug==3.0.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
